silence some more compiler warnings
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@10877 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ formats/format_pcm.c
@@ -199,7 +199,7 @@ static int pcm_seek(struct ast_filestream *fs, off_t sample_offset, int whence)
 		offset = sample_offset;
 	}
 	if (offset < 0) {
-		ast_log(LOG_WARNING, "negative offset %ld, resetting to 0\n", offset);
+		ast_log(LOG_WARNING, "negative offset %ld, resetting to 0\n", (long) offset);
 		offset = 0;
 	}
 	if (whence == SEEK_FORCECUR && offset > max) { /* extend the file */
@@ -215,7 +215,7 @@ static int pcm_seek(struct ast_filestream *fs, off_t sample_offset, int whence)
 		ret = 0; /* successful */
 	} else {
 		if (offset > max) {
-			ast_log(LOG_WARNING, "offset too large %ld, truncating to %ld\n", offset, max);
+			ast_log(LOG_WARNING, "offset too large %ld, truncating to %ld\n", (long) offset, (long) max);
 			offset = max;
 		}
 		ret = fseeko(fs->f, offset, SEEK_SET);
